+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.2.0
+* New event model
+* Cleaning up unnecessary libs
+* Performance improvements
+* Api changes:
+
+from:
+```
+serial.onSerialData.subscribe(
+```
+to:
+```
+serial.onSerialData.stream.listen(
+```
+
 ## 0.1.2
 * Android compilation